feat: Add password change functionality to user edit form
This commit adds the functionality to change the password in the user edit form. The code changes include adding the password input fields in the form-edit-user.handlebars view and updating the submitEditUser function in editUserForm.js to include the password in the payload when sending the PUT request to update the user. Additionally, the users.controller.js file has been updated to hash the password using bcrypt before updating the user data in the database.