Change password (#27)
* feat: Add password change functionality to user edit form

This commit adds the functionality to change the password in the user edit form. The code changes include adding the password input fields in the form-edit-user.handlebars view and updating the submitEditUser function in editUserForm.js to include the password in the payload when sending the PUT request to update the user. Additionally, the users.controller.js file has been updated to hash the password using bcrypt before updating the user data in the database.

* feat: Started email verification functionality to user registration

This commit adds the functionality to verify user email during the registration process. The code changes include generating a random verification code for each user in the UsersDTO class, updating the user.model.js schema to include the verifyCode field, and modifying the register route in sessions.api.js to send a verification email with the generated code. Additionally, the sessions.controller.js file has been updated to include a new verifyCode method that updates the user's verification status based on the provided code.

* Finished verify account flow

* feat: Add sessions API documentation

This commit adds the sessions API documentation to the Swagger documentation. It includes the endpoints for user registration, login, getting user data, signing out, and password reset. The code changes involve adding a new sessions.doc.yaml file with the OpenAPI specification for the sessions API, and updating the swagger.util.js file to include the sessions.doc.yaml file in the list of API files to be documented.

---------

Co-authored-by: MariaNavarro90 <navarromariavictoria8@gmail.com>