Documentacion (#31)
* chore: Update npm dependencies and Swagger documentation

Update the npm dependencies "stripe" from version 16.8.0 to 16.9.0 in the package.json and package-lock.json files. Also, update the Swagger documentation in swagger.util.js to include the new API files for ticket and payment. This commit improves the project's dependencies and keeps the documentation up to date.

* feat: Add carousel with welcome messages to landing page

This commit adds a carousel to the welcome.handlebars view on the landing page. The carousel displays three images with corresponding welcome messages. Each image is set to cover the entire viewport height. The carousel is interactive and can be navigated using the previous and next buttons. This enhancement improves the visual appeal and user experience of the landing page.

Co-authored-by: MariaNavarro90 <navarromariavictoria8@gmail.com>

* UI improvements and fixed slider issue that overlapped Login button on navbar

* More UI improvements